+++ src/odin/enhancedtrippath.cc
@@ -19,6 +19,7 @@ using namespace valhalla::baldr;
 
 namespace {
 constexpr float kShortRemainingDistanceThreshold = 0.402f; // Kilometers (~quarter mile)
+constexpr int kSignificantRoadClassThreshold = 2;          // Max lower road class delta
 
 const std::string& TripLeg_RoadClass_Name(int v) {
   static const std::unordered_map<int, std::string> values{
@@ -1357,6 +1358,25 @@ bool EnhancedTripLeg_Node::HasForwardTraversableIntersectingEdge(
   for (int i = 0; i < intersecting_edge_size(); ++i) {
     if (is_forward(GetTurnDegree(from_heading, intersecting_edge(i).begin_heading())) &&
         GetIntersectingEdge(i)->IsTraversableOutbound(travel_mode)) {
+      return true;
+    }
+  }
+  return false;
+}
+
+bool EnhancedTripLeg_Node::HasForwardTraversableSignificantRoadClassXEdge(
+    uint32_t from_heading,
+    const TripLeg_TravelMode travel_mode,
+    TripLeg_RoadClass path_road_class) {
+
+  for (int i = 0; i < intersecting_edge_size(); ++i) {
+    auto xedge = GetIntersectingEdge(i);
+    // if the intersecting edge is forward
+    // and is traversable based on mode
+    // and is a significant road class as compared to the path road class
+    if (is_forward(GetTurnDegree(from_heading, intersecting_edge(i).begin_heading())) &&
+        xedge->IsTraversableOutbound(travel_mode) &&
+        ((xedge->road_class() - path_road_class) <= kSignificantRoadClassThreshold)) {
       return true;
     }
   }